backend/pyproject.toml
[project]
name = "noema-backend"
version = "0.1.0"
description = "Noema — 팀용 지식 관리 도구 백엔드"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.0.0",
    "pgvector>=0.3.0",
    "neo4j>=5.0.0",
    "redis>=5.0.0",
    "anthropic>=0.40.0",
    "voyageai>=0.3.0",
    "sse-starlette>=2.0.0",
    "pyjwt>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
